hasProprty file added
