+++ server/controllers/loginController.js
@@ -1,15 +1,16 @@
 const db = require("../userModel");
 const User = db.user;
 const UserData = require("../userModel/userData");
+const hp = require("../utils/hasProperty");
 
 function loginUser(field){
     UserData.findDataByEmail(field.email,(result) => {
-        if(!hasErrorProperty(result)){
+        if(!hp.hasErrorProperty(result)){
             return { "status": "500", result };
         }
     
         console.log(result);
-        if(!hasUserProperty(result)){
+        if(!hp.hasUserProperty(result)){
             return { "status": "404", "error": "User Not found." };
         }
 
@@ -23,16 +24,6 @@ function loginUser(field){
 
         return {"token":token};
     });
-}
-
-function hasUserProperty(name){
-    console.log(name);
-    return name.hasOwnProperty('User');
-}
-
-function hasErrorProperty(name){
-    console.log(name);
-    return name.hasOwnProperty('error');
 }
 
 exports.loginUser = loginUser;
