+++ server/controllers/registerController.js
@@ -1,23 +1,20 @@
 const db = require("../userModel");
 const User = db.user;
 const UserData = require("../userModel/userData");
+const hp = require("../utils/hasProperty");
 
 function registerUser(field){
     const user = UserData.createNewUserObject(field);
     
     const result = UserData.saveDataToUserModel(user);
 
-    if(!hasValidProperty(result)){
+    if(!hp.hasValidProperty(result)){
         return { "status": "500", "error": result };;
     }
 
     const token = UserData.generateAuthToken(user);
 
     return {"token": token}; 
-}
-
-function hasValidProperty(name){
-    return name.hasOwnProperty('valid');
 }
 
 exports.registerUser = registerUser;
